+++ base/org/open2jam/parser/NoteEvent.java
@@ -4,7 +4,7 @@ public class NoteEvent extends Event
 {
 	protected short value;
 
-	public NoteEvent(double measure, short channel, short value)
+	public NoteEvent(double measure, int channel, short value)
 	{
 		this.measure = measure;
 		this.channel = channel;
